Milestone: Preserve asset source relationships
- Populate AudioAsset.SourceId from the originating Source
- Preserve source information during folder discovery
- Align AudioAsset.SourceId with Source.Id (long)
- Validate indexing pipeline and database persistence

+++ src/WemBam/Models/AudioAsset.cs
@@ -2,7 +2,7 @@ namespace WemBam.Models
 {
     public class AudioAsset
     {
-        public int SourceId { get; set; }
+        public long SourceId { get; set; }
 
         public string FileName { get; set; } = string.Empty;
